Allow autosave association methods to be overwritten
There are many rails gems out there that define active record classes. Even more applications consume those gems.

Sometimes, those applications want to override the properties of the rails gem active record classes. When you override an association by redefining it using a concern, prepended module, etc, rails generally lets you do this and respect it. For example, passing association arguments such as `disable_joins`, `foreign_key`, `inverse_of` to `has_many` in an overriding class will change the original ActiveRecord Class' behavior (e.g. association.rb:45:219:360)

However, if you try to override the `autosave` option, it does not change the behavior of the underlying class.

```
class OriginalClass < ActiveRecord::Base
  has_many :items
end

class MyExtension
    extend ActiveSupport::Concern

  # Adding autosave: true here doesn't do anything
  has_many :items, autosave: true

  OriginalClass.prepend(self)
end
```

This is because even though you can override the reflection attached to the class, the original reflection is still referenced within blocks that define the methods related to autosaving associations. This means that the original reflection behavior (in this example, autosave: false) is still used.

The `add_autosave_association_callbacks` method are only defined when an association is first built; or in the example I gave above, when an association is rebuilt. I was not able to get this method to fire more than the number of times association helpers like `has_many` were called in the app with a specific association name. I don't think that checking if the method was already defined was buying any performance improvements, but it was preventing the existing behavior on a active record class from being overwritten.